requirements.txt
openai-whisper==20231117
ffmpeg-python==0.2.0
click==8.1.7
tqdm==4.66.4